// src/types/employer/job-posting.ts
export type JobPostingType = {
  id: number;
  currency: string;
  jobType: string;
  jobTitle: string;
  number: string;
  skillAbilities: string;
  specificDuties: string;
  duty: string;
  startDate: Date;
  salary: string;
  hour: string;
  accommodation: string;
  visa: string;
  vacation: string;
  benefit: string;
  period: Date;
  dateRecruitment: Date;
  document: string;
  process: string;
  officer: string;
  other: string;
  status: number;
  isDelete: boolean;
  slug: null | string;
  statistical: any;
  apply: any;
  interview: any;
  country: any;
  position: any;
  periodStart: string;
  periodEnd: string;
  summary: string;
  language: any;
  insurance: string;
  positionName: string;
  company: any;
  totalClick: number;
  totalSave: number;
};